Backend: Add subcategories, setup redis, modularize schema, bug fixes

+++ backend/graphql/resolvers/category.js
@@ -1,28 +1,31 @@
 import Category from '../../models/categoryModel.js';
+import SubCategory from '../../models/subcategoryModel.js';
 import { loggedin, admin } from '../../utils/verifyUser.js';
 
+// Category
 const createCategory = async (args, { req, redis }) => {
   try {
-    // if(admin(req)) {
+    if (admin(req)) {
       const { name } = args;
 
       const resp = await Category.find({ name: name });
 
-      if(resp.length === 0) {
+      if (resp.length === 0) {
         const newCategory = new Category({
           name: name,
         });
-        await newCategory.save();
+        const res = newCategory.save();
+        return res;
+      } else {
+        throw new Error('Category already exists');
       }
-      
-      return { msg: 'success' };
-    // }
+    }
   } catch (err) {
     throw err;
   }
 };
 
-const categories = async () => {
+const getCategories = async (args, { req, redis }) => {
   try {
     const categories = await Category.find({});
     return categories;
@@ -33,7 +36,7 @@ const categories = async () => {
 
 const updateCategory = async (args, { req, redis }) => {
   try {
-    // if(admin(req)) {
+    if (admin(req)) {
       const { name, newName } = args;
 
       let updatedCategory = {
@@ -41,13 +44,10 @@ const updateCategory = async (args, { req, redis }) => {
       };
       updatedCategory = { $set: updatedCategory };
 
-      await Category.update(
-        { name: name },
-        updatedCategory
-      ).exec();
-      
+      await Category.update({ name: name }, updatedCategory).exec();
+
       return { msg: 'success' };
-    // }
+    }
   } catch (err) {
     throw err;
   }
@@ -55,16 +55,86 @@ const updateCategory = async (args, { req, redis }) => {
 
 const deleteCategory = async (args, { req, redis }) => {
   try {
-    // if(admin(req)) {
+    if (admin(req)) {
       const { name } = args;
 
       await Category.deleteOne({ name: name });
-      
+
       return { msg: 'success' };
-    // }
+    }
   } catch (err) {
     throw err;
   }
 };
 
-export { createCategory, categories, updateCategory, deleteCategory };
+// SubCategory
+const createSubCategory = async (args, { req, redis }) => {
+  try {
+    if (admin(req)) {
+      const { name, category } = args;
+
+      const resp = await SubCategory.find({ name: name, category: category });
+
+      if (resp.length === 0) {
+        const newSubCategory = new SubCategory({
+          name: name,
+          category: category,
+        });
+        const res = newSubCategory.save();
+        return res;
+      } else {
+        throw new Error('Subcategory already exists in given category');
+      }
+    }
+  } catch (err) {
+    throw err;
+  }
+};
+
+const getSubCategories = async (args, { req, redis }) => {
+  try {
+    const subCategories = await SubCategory.find({}).populate('category');
+    return subCategories;
+  } catch (err) {
+    throw err;
+  }
+};
+
+const updateSubCategory = async (args, { req, redis }) => {
+  try {
+    if (admin(req)) {
+      await SubCategory.update(
+        { _id: args.subCategoryId },
+        {
+          name: args.name,
+        }
+      ).exec();
+
+      return { msg: 'success' };
+    }
+  } catch (err) {
+    throw err;
+  }
+};
+
+const deleteSubCategory = async (args, { req, redis }) => {
+  try {
+    if (admin(req)) {
+      await SubCategory.deleteOne({ _id: args.subCategoryId });
+      return { msg: 'success' };
+    }
+  } catch (err) {
+    throw err;
+  }
+};
+
+export {
+  getCategories,
+  createCategory,
+  updateCategory,
+  deleteCategory,
+  getSubCategories,
+  createSubCategory,
+  updateSubCategory,
+  deleteSubCategory,
+};
